+++ client/src/components/CoasterCard.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import wagon from '../assets/coaster-wagon.svg';
-import { useState } from 'react';
 
 const Shadow = styled.div`
   filter: drop-shadow(2px 3px 3px rgba(50, 50, 0, 0.2));
@@ -59,35 +58,21 @@ const CardContent = styled.div`
 `;
 
 export default function CoasterCard({ image, name, ranking }) {
-  const [coasterInfos, setCoasterInfos] = useState([]);
-  async function getCoasterInfos() {
-    const response = await fetch('http://localhost:8080/coasters');
-    const newCoasters = await response.json();
-    setCoasterInfos(newCoasters);
-  }
-  React.useEffect(() => {
-    getCoasterInfos();
-  }, []);
-
   return (
-    <div>
-      {coasterInfos.map(info => (
-        <Shadow key={info.id}>
-          <Card>
-            <Image src={info.pic} />
-            <CardContent>
-              <Name>{info.name}</Name>
-              <Ranking>{info.ranking}</Ranking>
-              <RankingFix>/8</RankingFix>
-            </CardContent>
-            <CardFooter>
-              {[1, 2, 3, 4, 5, 6, 7, 8].map(score => (
-                <WagonImage key={score} src={wagon} highlight={score <= info.ranking} />
-              ))}
-            </CardFooter>
-          </Card>
-        </Shadow>
-      ))}
-    </div>
+    <Shadow>
+      <Card>
+        <Image src={image} />
+        <CardContent>
+          <Name>{name}</Name>
+          <Ranking>{ranking}</Ranking>
+          <RankingFix>/8</RankingFix>
+        </CardContent>
+        <CardFooter>
+          {[1, 2, 3, 4, 5, 6, 7, 8].map(score => (
+            <WagonImage key={score} src={wagon} highlight={score <= ranking} />
+          ))}
+        </CardFooter>
+      </Card>
+    </Shadow>
   );
 }
