+++ db.json
@@ -8,7 +8,7 @@
       "length": "1320 m",
       "duration": "100 sec.",
       "hight": "30 m",
-      "pic": "https://upload.wikimedia.org/wikipedia/commons/thumb/2/2c/Taron-1.jpg/1280px-Taron-1.jpg"
+      "pic": "https://i.pinimg.com/originals/6b/fb/27/6bfb27f7974394e504d2197824109a85.jpg"
     },
     {
       "id": 2,
